pyproject.toml
[project]
name = "autogenanalyzergpt"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autogen-agentchat>=0.6.4",
    "autogen-core>=0.6.4",
    "autogen-ext[docker,langchain,openai]>=0.6.4",
    "autogenstudio>=0.1.5",
    "ddgs>=9.4.3",
    "ipykernel>=6.30.0",
    "langchain-community>=0.3.21",
    "pandas>=2.3.1",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "streamlit>=1.47.1",
]
